Update README.md
The Decorator design pattern is a structural pattern in software design that allows you to dynamically add new behaviors or responsibilities to objects without altering their existing code. It involves creating a set of decorator classes that wrap concrete components and add additional functionality to them. This pattern promotes flexibility, code reusability, and separation of concerns while adhering to the Open-Closed Principle, making it useful for building extensible and maintainable systems.